% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdriver_package.R
\name{add_log}
\alias{add_log}
\title{Create a log entry to stdio and/or a Logdriver server}
\usage{
add_log(
  level = c("info", "warn", "error", "critical"),
  username = NA,
  event = NA,
  description = NA,
  logdriver_appname = NA,
  logdriver_host = NA,
  logdriver_port = NA,
  logdriver_apikey = NA
)
}
\arguments{
\item{level}{Log level. Values: "info" (default), "warn", "error", "critical".}

\item{username}{Character.}

\item{event}{Character.}

\item{description}{Character.}

\item{logdriver_appname}{Mandatory for Logdriver server. Overrides environment variable LOGDRIVER_APPNAME.}

\item{logdriver_host}{Mandatory for Logdriver server. Overrides environment variable LOGDRIVER_HOST}

\item{logdriver_port}{Mandatory for Logdriver server. Overrides environment variable LOGDRIVER_PORT}

\item{logdriver_apikey}{Placeholder for future use.}
}
\value{

}
\description{
Create a log entry to stdio and/or a Logdriver server
}
\examples{
\dontrun{
# write log to console
add_log(level = "info", username = "chris", event = "created log", description = "this is a really informative log printed to stdout")

# write log to Logdriver server hosted remotely
add_log(level = "info", username = "chris", event = "created log", description = "this log is going on a logdriver server", logdriver_appname = "testapp", logdriver_host = "logdriver-test.fly.dev", logdriver_port = "8000" )
}
}
